Edited lab4 simple_interest_calculator.py

+++ module01/lab4-simple-interest-calculator/simple_interest_calculator.py
@@ -21,7 +21,7 @@ while True:
 # Gather time from user and validate
 while True:
     try:
-        time = int(input("Enter time (years): "))
+        time = float(input("Enter time (years): "))
         if time > 0:
             break
         else:
